[libc] Fix some sign-compare warnings

+++ libc/error/error.c
@@ -13,7 +13,7 @@ strerror(int err)
 {
    int fd = -1;
    int cc;
-   int bufoff = 0;
+   size_t bufoff = 0;
    char inbuf[256];
    static char retbuf[60];
